fixed tasks
