+++ src/App.jsx
@@ -1,13 +1,13 @@
-import { Header } from "./components/Header";
-import "./App.css";
-import { Context } from "./helpers/Context";
+import { Header } from './components/Header';
+import './App.css';
+import { Context } from './components/contexts/Context';
 
 function App() {
-  return (
-    <Context>
-      <Header />
-    </Context>
-  );
+    return (
+        <Context>
+            <Header />
+        </Context>
+    );
 }
 
 export default App;
