+++ src/components/Header.jsx
@@ -1,31 +1,31 @@
-import icon from "../assets/images/night.png";
-import style from "./Header.module.scss";
-import iconMoon from "../assets/images/moon.png";
-import { useContext } from "react";
-import { CustomContext } from "../helpers/Context";
+import style from './Header.module.scss';
+import iconMoonNight from '../assets/images/half-moon-shape-svgrepo-com.svg';
+import { useContext } from 'react';
+import { ThemContext } from './contexts/Context';
+import iconMoon from '../assets/images/crescent-moon-phase-svgrepo-com.svg';
 
 function Header() {
-  const { isTheme = true, toggleTheme } = useContext(CustomContext);
-  return (
-    <>
-      {isTheme ? (
-        <div className={style.head_block}>
-          <h1>Where in the world?</h1>
-          <h3 onClick={toggleTheme}>
-            <img src={icon} alt="moon"></img>
-            Dark Mode
-          </h3>
-        </div>
-      ) : (
-        <div className={`${style.head_block} ${style.light}`}>
-          <h1>Where in the world?</h1>
-          <h3 onClick={toggleTheme}>
-            <img src={iconMoon} alt="iconMoon"></img>
-            Light Mode
-          </h3>
-        </div>
-      )}
-    </>
-  );
+    const { isTheme = 'dark', toggleTheme } = useContext(ThemContext);
+    return (
+        <>
+            {isTheme === 'dark' ? (
+                <div className={style.head_block}>
+                    <h1>Where in the world?</h1>
+                    <h2 onClick={toggleTheme}>
+                        <img src={iconMoonNight} alt="moon" />
+                        Dark Mode
+                    </h2>
+                </div>
+            ) : (
+                <div className={`${style.head_block} ${style.light}`}>
+                    <h1>Where in the world?</h1>
+                    <h2 onClick={toggleTheme}>
+                        <img src={iconMoon} alt="iconMoon" />
+                        Light Mode
+                    </h2>
+                </div>
+            )}
+        </>
+    );
 }
 export { Header };
